Fix not releasing adc unit in image wave

+++ src/graphics3d.h
@@ -4,7 +4,7 @@
 typedef struct { int16_t x; int16_t y;} vec2;
 typedef struct { float x; float y; float z;} vec3f;
 typedef struct { float x; float y;} vec2f;
-typedef struct {uint16_t r; uint16_t g; uint16_t b;} colourtype;
+typedef struct {uint8_t r; uint8_t g; uint8_t b;} colourtype;
 
 
 static inline vec3f sub3d(const vec3f p0, const vec3f p1) {
